moves types to '@frontity/tiny-router/types'
