+++ packages/tiny-router/src/index.ts
@@ -1,4 +1,4 @@
-import TinyRouter from "..";
+import TinyRouter from "../types";
 import { set, init, beforeSSR } from "./actions";
 
 const tinyRouter: TinyRouter = {
